# docs/04-api/openapi.yaml
openapi: 3.1.0
info:
  title: Personal Intelligence Layer API
  version: 0.1.0
  description: |
    REST API for PIL. Используется Admin UI и опционально внешними MCP-клиентами через REST-фасад.
    Полная семантика MCP-инструментов — docs/04-api/mcp-api.md.
servers:
  - url: https://pil.local/api/v1
    description: Self-host через reverse-proxy
  - url: http://localhost:8080/v1
    description: Локальная разработка
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /persons/{id}/context:
    get:
      summary: Composed context (persona + tasks + interactions + graph + memory)
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
        - { in: query, name: query, schema: { type: string } }
        - { in: query, name: hops, schema: { type: integer, default: 1, maximum: 3 } }
        - { in: query, name: k_memory, schema: { type: integer, default: 5, maximum: 50 } }
      responses:
        '200':
          description: OK
          content:
            application/json: { schema: { $ref: '#/components/schemas/PersonContext' } }
        '206':
          description: Partial (один из бэкендов недоступен)
          content:
            application/json: { schema: { $ref: '#/components/schemas/PersonContext' } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: pil_session
...
  schemas:
...
    Person:
      type: object
      properties:
        id:               { type: string, format: uuid }
        tg_user_id:       { type: integer, nullable: true }
        username:         { type: string, nullable: true }
        display_name:     { type: string }
        role:             { type: string, nullable: true }
        organizations:    { type: array, items: { type: string } }
        topics:           { type: array, items: { type: string } }
        communication_style: { type: string, nullable: true }
        trust_score:      { type: number, minimum: 0, maximum: 1 }
        last_interaction_at: { type: string, format: date-time, nullable: true }
        notes:            { type: string, nullable: true }
...
    PersonContext:
      type: object
      properties:
        person: { $ref: '#/components/schemas/Person' }
        graph:
          type: object
          properties:
            neighbors: { type: array, items: { $ref: '#/components/schemas/Person' } }
            chats_in_common: { type: array, items: { type: object } }
        open_tasks: { type: array, items: { $ref: '#/components/schemas/Task' } }
        recent_interactions: { type: array, items: { $ref: '#/components/schemas/Interaction' } }
        memory: { type: array, items: { $ref: '#/components/schemas/MemoryChunk' } }
        context_markdown: { type: string }
        partial_reasons: { type: array, items: { type: string }, nullable: true }
    Interaction:
      type: object
      properties:
        id:           { type: string, format: uuid }
        chat_id:      { type: string, format: uuid }
        window_start: { type: string, format: date-time }
        window_end:   { type: string, format: date-time }
        topics:       { type: array, items: { type: string } }
        sentiment:    { type: string, enum: [positive, negative, neutral, mixed] }
        summary:      { type: string }
    MemoryChunk:
      type: object
      properties:
        chunk_id: { type: string, format: uuid }
        type:     { type: string, enum: [summary, note, quote, fact] }
        text:     { type: string }
        person_ids: { type: array, items: { type: string, format: uuid } }
        chat_id:  { type: string, format: uuid, nullable: true }
        created_at: { type: string, format: date-time }
    Task:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        description: { type: string, nullable: true }
        owner_person_id: { type: string, format: uuid, nullable: true }
        counterpart_person_id: { type: string, format: uuid, nullable: true }
        chat_id: { type: string, format: uuid, nullable: true }
        due_at: { type: string, format: date-time, nullable: true }
        status: { type: string, enum: [open, done, dropped] }
        priority: { type: integer, minimum: 1, maximum: 5 }
        confidence: { type: number, minimum: 0, maximum: 1 }
...